move search page to /event_series and move event series create & edit pages to /workshop
