+++ app/workshop/[id]/add_event_card_container.tsx
