+++ app/workshop/[id]/add_event_to_series.tsx
