+++ app/workshop/[id]/edit_series_container.tsx
