+++ app/workshop/[id]/edit_series_content.tsx
