+++ app/workshop/[id]/edit_series_events.tsx
