+++ app/workshop/[id]/edit_series_tabs.tsx
