+++ app/workshop/[id]/page.tsx
