+++ app/workshop/components/delete_event_series_form.tsx
