+++ app/workshop/components/events_list.tsx
