+++ app/workshop/create/page.tsx
@@ -1,6 +1,6 @@
 import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 import LoginPrompt from '@/app/components/login_prompt';
-import AddEventSeriesForm from '@/app/event_series/components/add_event_series_form';
+import AddEventSeriesForm from '@/app/workshop/components/add_event_series_form';
 import { getServerSession } from 'next-auth';
 import React from 'react';
